requirements.txt
python-dotenv
langchain_google_genai
langchain_community
ipykernel
bs4
arxivloader
pymupdf
arxiv
wikipedia
langchain-text-splitters
chromadb
notebook
jupyter
langchain-ollama
langchain_huggingface
sentence-transformers
faiss-cpu
langchain_chroma
streamlit
langchain_groq
langchain_huggingface
duckdb
pandas
duckduckgo_search==5.3.1b1
mysql-connector-python
SQLAlchemy
validators
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
fastapi
langserve
sse_starlette
pydantic
uvicorn
duckduckgo-search
tenacity
gradio
pymongo
sqlalchemy
pymysql
passlib[bcrypt]
python-jose
transformers
pypdf2
langchain_nvidia_ai_endpoints
crewai
crewai_tools
tools
yt_dlp==2023.11.14
youtube-transcript-api==0.6.1
phidata
yfinance
packaging
duckduckgo-search
groq
pgvector
psycopg[binary]
duckduckgo-search
agno
rich
tavily-python==0.7.6
requests==2.32.3
typing-extensions==4.12.2
setuptools==70.0.0
starlette>=0.40.0
langgraph-prebuilt==0.2.2
langchain-tavily==0.2.6
browser_use
playwright
litellm
smolagents
ddgs #for DuckDuckGOSearch
matplotlib
bytez